Feature/#23 create navigation bar (#24)
* Edit global.css and index.html for working navigational bar

Uses the images in assets.

* Get rid of lego.css and lego.html

They are not needed in the example that Clare made.

* Change to text rather than images for header content

Text allows for better css and won't annoy us in the future. I could find the font "Spartacus", but not "Segoe UI", although I allowed the code to use it if it is installed locally. I have the backup of a similar font from google.

* Change color to gray (0x707070) instead of dark gray (0x202020)

* added hamburger menu for width smaller than 1240px

* set correct colors for hamburger icon

---------

Co-authored-by: HarlanHaller <harlanhaller@icloud.com>