Feature/add core values (#27)
* Edit global.css and index.html for working navigational bar

Uses the images in assets.

* Change to text rather than images for header content

Text allows for better css and won't annoy us in the future. I could find the font "Spartacus", but not "Segoe UI", although I allowed the code to use it if it is installed locally. I have the backup of a similar font from google.

* Change color to gray (0x707070) instead of dark gray (0x202020)

* Add Core Values content

Includes a 2x3 grid with each item a box with a background gradient, and text on top.